Denmark Metro station and Real estate data EDA code
This project is quite big and this resposity is only the small part of whole project. This resp. correspond to geostatial analysis combining real estate listing and Metro station data. The task of the project: identify the impact of new new built metro stations to changes in prices of 500 m around real estate properties.